some reorganisation of code
